1.7.2

+++ src/com/watabou/pixeldungeon/levels/features/HighGrass.java
@@ -1,5 +1,4 @@
 /*
- * Pixel Dungeon
  * Copyright (C) 2012-2014  Oleg Dolya
  *
  * This program is free software: you can redistribute it and/or modify
@@ -17,6 +16,7 @@
  */
 package com.watabou.pixeldungeon.levels.features;
 
+import com.watabou.pixeldungeon.Challenges;
 import com.watabou.pixeldungeon.Dungeon;
 import com.watabou.pixeldungeon.actors.Char;
 import com.watabou.pixeldungeon.actors.buffs.Barkskin;
@@ -40,22 +40,24 @@ public class HighGrass {
 		Level.set( pos, Terrain.GRASS );
 		GameScene.updateMap( pos );
 		
-		int herbalismLevel = 0;
-		if (ch != null) {
-			Herbalism herbalism = ch.buff( Herbalism.class );
-			if (herbalism != null) {
-				herbalismLevel = herbalism.level;
+		if (!Dungeon.isChallenged( Challenges.NO_HERBALISM )) {
+			int herbalismLevel = 0;
+			if (ch != null) {
+				Herbalism herbalism = ch.buff( Herbalism.class );
+				if (herbalism != null) {
+					herbalismLevel = herbalism.level;
+				}
 			}
-		}
-		
-		// Seed
-		if (herbalismLevel >= 0 && Random.Int( 18 ) <= Random.Int( herbalismLevel + 1 )) {
-			level.drop( Generator.random( Generator.Category.SEED ), pos ).sprite.drop();
-		}
-		
-		// Dew
-		if (herbalismLevel >= 0 && Random.Int( 6 ) <= Random.Int( herbalismLevel + 1 )) {
-			level.drop( new Dewdrop(), pos ).sprite.drop();
+			
+			// Seed
+			if (herbalismLevel >= 0 && Random.Int( 18 ) <= Random.Int( herbalismLevel + 1 )) {
+				level.drop( Generator.random( Generator.Category.SEED ), pos ).sprite.drop();
+			}
+			
+			// Dew
+			if (herbalismLevel >= 0 && Random.Int( 6 ) <= Random.Int( herbalismLevel + 1 )) {
+				level.drop( new Dewdrop(), pos ).sprite.drop();
+			}
 		}
 		
 		int leaves = 4;
